+++ storybook/cypress/e2e/scrollable.cy.ts
@@ -53,6 +53,7 @@ describe("Scrollable", () => {
       .should("not.have.class", "dso-scroll-top");
   });
 
+  // enable via #3775
   it.skip("should update scroll state with dynamic content", () => {
     cy.visit("http://localhost:45000/iframe.html?args=&id=core-scrollable--dynamic-content");
     const scrollContainer = () => cy.get("dso-scrollable.hydrated").shadow().find(".dso-scroll-container");