+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 * List Button: Voorlezen bij focus werkt niet ([#3811](https://github.com/dso-toolkit/dso-toolkit/issues/3811))
+* Search Bar: Focus verdwijnt bij TAB vanuit text input van Autosuggest ([#3813](https://github.com/dso-toolkit/dso-toolkit/issues/3813))
 
 ## 🚑 Release 99.1.0 - 2026-07-13